+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Displaying sandboxed `ScoreCard` data.
 
 ## `20230623t075300-all`
  * Bumped runtimeVersion to `2023.06.21`.